chore: add validations and find/get to meetings module

+++ src/modules/meetings/services/DeleteMeetingMemberService.v1.ts
@@ -1,3 +1,4 @@
+import { AppError } from "@/shared/errors/AppError";
 import { prisma } from "@/shared/infra/prisma/connection";
 
 type DeleteMeetingMemberV1 = {
@@ -10,6 +11,18 @@ class DeleteMeetingMemberV1Service {
     meetingId,
     memberId,
   }: DeleteMeetingMemberV1): Promise<void> {
+    const memberAlreadyExists = await prisma.meetingMember.findFirst({
+      where: { meetingId, memberId },
+    });
+
+    if (!memberAlreadyExists) {
+      throw new AppError({
+        message: "Member is not assigned to this meeting",
+        code: "demolay.member-not-assigned",
+        statusCode: 404,
+      });
+    }
+
     await prisma.meetingMember.delete({
       where: { memberId_meetingId: { meetingId, memberId } },
     });
